[refactor] style ts 파일 제거
