+++ src/Utils/constants.ts
@@ -1,0 +1,9 @@
+export const PAGE_TITLE = {
+  board: 'Boards',
+  members: 'Members',
+};
+
+export const FOOTER = {
+  location: '창원대학교 51호관 113호',
+  copyRight: '2023. Casper All rights reserved.',
+};
