+++ src/components/common/Button/index.tsx
@@ -1,6 +1,13 @@
+import { ButtonHTMLAttributes } from 'react';
 import styled from 'styled-components';
 
-export const ButtonWrapper = styled.button`
+interface Props extends ButtonHTMLAttributes<HTMLButtonElement> {}
+
+function Button({ ...props }: Props) {
+  return <ButtonWrapper {...props} />;
+}
+
+const ButtonWrapper = styled.button`
   width: 90px;
   height: 40px;
   color: ${({ theme }) => theme.white};
@@ -12,3 +19,5 @@ export const ButtonWrapper = styled.button`
     background-color: ${({ theme }) => theme.green200};
   }
 `;
+
+export default Button;
