+++ src/pages/adminpage/index.tsx
@@ -1,4 +1,4 @@
-import { Wrapper, SideBar, SideBarRow, SideTitle } from './adminpage.style';
+import styled from 'styled-components';
 
 export default function Login() {
   return (
@@ -15,3 +15,33 @@ export default function Login() {
     </Wrapper>
   );
 }
+
+const Wrapper = styled.div`
+  display: flex;
+  align-items: center;
+  position: absolute;
+  width: 100%;
+  height: 90%;
+`;
+const SideBar = styled.div`
+  display: flex;
+  align-items: center;
+  position: absolute;
+  width: 30%;
+  height: 100%;
+  background-color: ${({ theme }) => theme.borderDefault};
+  overflow: scroll;
+  // justify-content:center;
+  flex-direction: column;
+`;
+const SideBarRow = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  border: solid 2px #808e9b;
+  width: 100%;
+`;
+const SideTitle = styled.h1`
+  padding: 0.7em;
+`;
+const OpenMenu = styled.div``;
